requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==45.0.4
cssselect2==0.8.0
Django==5.2.3
django-appconf==1.1.0
django-celery-beat==2.8.1
django-imagekit==5.0.0
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
django-timezone-field==7.1
django_celery_results==2.6.0
dotenv==0.9.9
et_xmlfile==2.0.0
Flask==3.1.1
frozenlist==1.7.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
lxml==5.4.0
MarkupSafe==3.0.2
multidict==6.6.3
mysqlclient==2.2.7
numpy==2.3.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.0
phonenumbers==9.0.7
pilkit==3.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
pycparser==2.22
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pypdf==5.6.0
python-bidi==0.6.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
twilio==9.6.4
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
yarl==1.20.1
